Include LBI calculations.
Move all LBI work to tree_LBI.py and spin off of json/dendropy
conversions to tree_util.
