+++ augur/date_util.py
@@ -24,6 +24,17 @@ def numerical_date(date):
 def string_to_numerical_date(string):
 	date = string_to_date(string)
 	return numerical_date(date)
+	
+def date_to_day(date):
+    if isinstance(date, datetime.date):
+        return date.toordinal()
+    elif isinstance(date,basestring):
+        return datetime.datetime.strptime(date, '%Y-%m-%d').toordinal()
+    elif isinstance(date, int):
+        return date
+    else:
+        print "unknown date format", date
+        return np.nan	
 
 def main():
 	date1 = string_to_date('2011-06-12')
