+++ augur/streamline.py
@@ -8,7 +8,7 @@ def main():
 	print "--- Streamline at " + time.strftime("%H:%M:%S") + " ---"	
     
     # Streamline tree for auspice
-	tree = read_json('data/tree_refine.json')
+	tree = read_json('data/tree_LBI.json')
 	for node in all_descendants(tree):
 		node.pop("seq", None)
 		node.pop("clade", None)
